requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
aniso8601==7.0.0
asgiref==3.5.0
async-timeout==4.0.2
attrs==21.4.0
autobahn==22.2.2
Automat==20.2.0
cffi==1.15.0
channels==3.0.4
charset-normalizer==2.0.12
click==8.0.4
constantly==15.1.0
cryptography==36.0.1
daphne==3.0.2
Django==3.0.8
django-channels-graphql-ws==0.9.1
frozenlist==1.3.0
graphene==2.1.9
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
gunicorn==20.0.4
h11==0.13.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
msgpack==1.0.3
multidict==6.0.2
promise==2.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==22.0.0
pytz==2021.3
Rx==1.6.1
service-identity==21.1.0
singledispatch==3.7.0
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
Twisted==22.1.0
txaio==22.2.1
typing_extensions==4.1.1
uvicorn==0.17.5
yarl==1.7.2
zope.interface==5.4.0
